增加屏蔽词管理页面

+++ admin/src/main/java/jforgame/admin/system/dao/SysDictDao.java
@@ -3,7 +3,7 @@ package jforgame.admin.system.dao;
 import jforgame.admin.domain.SysDict;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SysDictDao extends JpaRepository<SysDict, Long> {
+public interface SysDictDao extends JpaRepository<SysDict, Integer> {
 
 
 }
